+++ CHANGELOG.md
@@ -2,7 +2,9 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-_(none)_
+* Upgrade to v1.96.0 of the AliCloud Terraform Provider
+* Upgrade to pulumi-terraform-bridge v2.8.0
+* Upgrade to Pulumi v2.10.0
 
 ---